feat(power): extract the PowerView into its own file.
